requirements.txt
fastapi==0.54.1
uvicorn==0.11.3
motor==2.0.0
pytest==5.3.0
telethon==1.13.0
python-dotenv==0.13.0
hachoir==3.1.1
Pillow==7.1.2